@brandonhackney 2024-01-23
I'm specifically drawing attention to the fact that [commit 7ae4a3a](https://github.com/VPNLirvine/trEYEcopa/pull/15/commits/7ae4a3a8b4e5b393d783981db7a80b430bd569d0) turns on the 'SkipSyncTests' flag, regardless of whether the experiment is run in debug mode or not. 

But we're rippin on ahead anyway